ar: Require ReaderAt, provide SectionReader
This is a breaking change: it drops support for non-seekable file
streams and file-like interfaces that do not provide the ReaderAt
interface.

In exchange for this loss of functionality, the `ar.Data` interfaces
provided are now themselves seekable and rereadable, and still valid
even after a subsequent portion of the same ar file has been consumed.
Similarly, it allows skipping over file content that does not need to be
read.