Rewrite the Dependency Architecture code
My old code was very annoying -- it was worse than annoying, it was
*wrong*. It totally didn't deal with the libc in the tuple -- and it
didn't use the correct method to match the tuples - the dpkg tupletable.

I vendored a copy of the Debian tupletable since I don't want to have it
require a Debian host, but I kept the same format so it could be
copy-pasted when it changes.

This is also annoying because any existing code using Arch {} directly
is likely to break -- however, the code was wrong anyway, so, uh,
apologies in advance.